Move NestedStoragePath to preserver namespace.
