+++ app/services/preserver/blind_importer/file_metadata_adapter/query_service/blind_importer_metadata_wayfinder.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 class Preserver::BlindImporter::FileMetadataAdapter::QueryService
   # Wayfinder so that FileMetadataResources can return a parent used by
-  # NestedStoragePath, as a way to figure out where a JSON or binary file would
+  # Preserver::NestedStoragePath, as a way to figure out where a JSON or binary file would
   # be stored in a nested hierarchy.
   class BlindImporterMetadataWayfinder < BaseWayfinder
     def parent
