+++ app/services/preserver/blind_importer/file_metadata_adapter/query_service/file_metadata_resource.rb
@@ -3,7 +3,7 @@ class Preserver::BlindImporter::FileMetadataAdapter::QueryService
   # Class used as an adapter for generating paths to be used in
   # Preserver::BlindImporter::FileMetadataAdapter::QueryService, so that
   # Preserver::BlindImporter::BlindImporterMetadataWayfinder can find parents to
-  # be used in NestedStoragePath.
+  # be used in Preserver::NestedStoragePath.
   class FileMetadataResource < Valkyrie::Resource
     attribute :ancestors
   end
